Abandoning the windows plugin
Considering all the things going on behind the scenes to position UI windows, a plugin that actually exposes all these things through plugin parameters would be both a nightmare to develop and probably a nightmare for a user to actually utilize. They might as well just learn javascript at that point.